框架搭建

+++ src/main/java/com/wei/config/shiro/ShiroUser.java
@@ -17,7 +17,7 @@ import java.util.Set;
 public class ShiroUser implements Serializable {
     private static final long serialVersionUID = -1373760761780840081L;
     
-    private Long id;
+    private Integer id;
     private final String loginName;
     private String name;
     private Set<String> urlSet;
@@ -27,7 +27,7 @@ public class ShiroUser implements Serializable {
         this.loginName = loginName;
     }
 
-    public ShiroUser(Long id, String loginName, String name, Set<String> urlSet) {
+    public ShiroUser(Integer id, String loginName, String name, Set<String> urlSet) {
         this.id = id;
         this.loginName = loginName;
         this.name = name;
